feat(sites): add cached repoInfo and add lastUpdated to sites route
